+++ LojaDeGames/src/main/java/com/minhaLojaDeGames/controller/ProdutoController.java
@@ -7,6 +7,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,12 +19,14 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.minhaLojaDeGames.model.Produto;
 import com.minhaLojaDeGames.repository.ProdutoRepository;
+import com.minhaLojaDeGames.service.ProdutoService;
 
 @RestController
 @RequestMapping("produto")
 public class ProdutoController {
 
 	private @Autowired ProdutoRepository repository;
+	private @Autowired ProdutoService service;
 	
 	@GetMapping("/todos")
 	public List<Produto> findAllProdutos() {
@@ -58,11 +61,22 @@ public class ProdutoController {
 	
 	@PutMapping("/put")
 	public ResponseEntity<Produto> putProduto(@Valid @RequestBody Produto produtoParaAtualizar) {
-		return ResponseEntity.status(201).body(repository.save(produtoParaAtualizar));
+		Optional<?> objetoProduto = service.atualizarProduto(produtoParaAtualizar);
+		if(objetoProduto.isEmpty()) {
+			return ResponseEntity.status(400).build();
+		}else {
+			return ResponseEntity.status(201).body(repository.save(produtoParaAtualizar));
+		}
 	}
 
 	@DeleteMapping("/delete/{id_produto}")
-	public void deleteProduto(@PathVariable(value = "id_produto") Long idProduto) {
-		repository.deleteById(idProduto);
+	public ResponseEntity<Produto> deleteProduto(@PathVariable(value = "id_produto") Long idProduto) {
+		Optional<Produto> objetoProduto = repository.findById(idProduto);
+		if(objetoProduto.isPresent()) {
+			repository.deleteById(idProduto);
+			return ResponseEntity.status(200).build();
+		}else {
+			return ResponseEntity.status(400).build();
+		}
 	}
 }
